Cleanup and restructuring
